requirements.txt
openai
flask
python-dotenv
flask-cors